Step 5+ : For loop is now better
For loop does not execute its body if the condition is false upon entry.

For loop returns the counter value upon termination, which is the value
that made it break the loop.

For loop looks more like the classical C-loop, except for the last part.
For example to loop from 1 to 10:

for( i = 1, i < 11, i + 1) loopbody